fix: StatisticsService 삭제
StatisticsService 삭제

- 기존 StatisticsService 클래스 내에 있던 calculateWinningStatistics 메서드 -> LottoService로 이동
- LottoController의 인스턴스 4개 -> 3개로 감소
- 현재 서비스 로직에서 담당하고 있는 역할이 적은 관계로 과한 서비스 분리라고 판단함
